import match status correctly
