+++ backend/app/schedulers/footballApi/util.ts
@@ -1,4 +1,4 @@
-import { Match, MatchStatus } from "../../../db/models/match.model";
+import { Match, getMatchStatus } from "../../../db/models/match.model";
 
 export const matchChanged = (apiMatch: any, dbMatch: Match) => {
   if (!dbMatch) {
@@ -36,20 +36,4 @@ export const makeMatchUpdate = (apiMatch: any) => {
   };
   const matchStatus = getMatchStatus(status);
   return { result, status: matchStatus, odds }
-}
-
-export const getMatchStatus = (status: string) => {
-  const MATCH_STATUS: { [key: string]: MatchStatus } = {
-    'SCHEDULED': MatchStatus.SCHEDULED,
-    'TIMED': MatchStatus.SCHEDULED,
-    'IN_PLAY': MatchStatus.LIVE,
-    'PAUSED': MatchStatus.LIVE,
-    'FINISHED': MatchStatus.FINISHED,
-    'AWARDED': MatchStatus.FINISHED,
-    'CANCELLED': MatchStatus.CANCELED,
-    'SUSPENDED': MatchStatus.CANCELED,
-    'POSTPONED': MatchStatus.POSTPONED
-  };
-
-  return MATCH_STATUS[status] || MatchStatus.SCHEDULED;
 }
